moved generic.py to shared - backend_core version is kept unchanged for now
